Move ini_to_dict from main script to segments.py
